Change envs directory
