+++ src/envs/__init__.py
@@ -1,5 +1,5 @@
 """
-Registration for Half-cheetah environments
+Registration code for Half-cheetah environments
 """
 
 import importlib
@@ -26,4 +26,4 @@ def register_env(name):
 for file in os.listdir(os.path.dirname(__file__)):
     if file.endswith(".py") and not file.startswith("_"):
         module = file[: file.find(".py")]
-        importlib.import_module("src.pearl.envs." + module)
+        importlib.import_module("src.envs." + module)
